+++ demo/api/src/config/environment-variables.ts
@@ -96,11 +96,10 @@ export class EnvironmentVariables {
     @IsString()
     S3_BUCKET: string;
 
-<<<<<<< HEAD
     @IsString()
     @ValidateIf(() => process.env.NODE_ENV === "production")
     CDN_ORIGIN_CHECK_SECRET: string;
-=======
+
     @ValidateIf((v) => v.AZURE_OPEN_AI_CONTENT_GENERATION_API_KEY || v.AZURE_OPEN_AI_CONTENT_GENERATION_DEPLOYMENT_ID)
     @IsString()
     AZURE_OPEN_AI_CONTENT_GENERATION_API_URL?: string;
@@ -112,5 +111,4 @@ export class EnvironmentVariables {
     @ValidateIf((v) => v.AZURE_OPEN_AI_CONTENT_GENERATION_API_URL || v.AZURE_OPEN_AI_CONTENT_GENERATION_API_KEY)
     @IsString()
     AZURE_OPEN_AI_CONTENT_GENERATION_DEPLOYMENT_ID?: string;
->>>>>>> main
 }