+++ packages/site/cms-site/src/blocks/DamFileDownloadLinkBlock.tsx
@@ -16,13 +16,6 @@ export const DamFileDownloadLinkBlock = withPreview(
             return children;
         }
 
-<<<<<<< HEAD
-        if (openFileType === "Download") {
-            return React.cloneElement(children, { href: file.fileUrl, title: title });
-        } else {
-            return React.cloneElement(children, { href: file.fileUrl, target: "_blank", rel: "noreferrer", title: title });
-        }
-=======
         const childProps = {
             href: file.fileUrl,
             title,
@@ -30,7 +23,6 @@ export const DamFileDownloadLinkBlock = withPreview(
         };
 
         return React.cloneElement(children, childProps);
->>>>>>> main
     },
     { label: "DamFileDownloadLink" },
 );