+++ demo/api/src/app.module.ts
@@ -1,10 +1,12 @@
 import {
     AccessLogModule,
+    AzureOpenAiContentGenerationModule,
     BlobStorageModule,
     BLOCKS_MODULE_TRANSFORMER_DEPENDENCIES,
     BlocksModule,
     BlocksTransformerMiddlewareFactory,
     BuildsModule,
+    ContentGenerationModule,
     ContentScope,
     CronJobsModule,
     DamModule,
@@ -23,6 +25,7 @@ import { DynamicModule, Module } from "@nestjs/common";
 import { Enhancer, GraphQLModule } from "@nestjs/graphql";
 import { Config } from "@src/config/config";
 import { ConfigModule } from "@src/config/config.module";
+import { ContentGenerationService } from "@src/content-generation/content-generation.service";
 import { DbModule } from "@src/db/db.module";
 import { LinksModule } from "@src/links/links.module";
 import { PagesModule } from "@src/pages/pages.module";
@@ -156,6 +159,14 @@ export class AppModule {
                     directory: `${config.blob.storageDirectoryPrefix}-public-uploads`,
                     acceptedMimeTypes: ["application/pdf", "application/x-zip-compressed", "application/zip"],
                 }),
+                ...(config.contentGeneration
+                    ? [
+                          ContentGenerationModule.register({
+                              Service: ContentGenerationService,
+                              imports: [AzureOpenAiContentGenerationModule.register(config.contentGeneration)],
+                          }),
+                      ]
+                    : []),
                 NewsModule,
                 MenusModule,
                 FooterModule,