+++ packages/admin/cms-admin/src/blocks/DamVideoBlock.tsx
@@ -1,6 +1,6 @@
-import { gql } from "@apollo/client";
-import { Field, FieldContainer, FinalFormSwitch, messages } from "@comet/admin";
-import { Delete, Video } from "@comet/admin-icons";
+import { gql, useApolloClient } from "@apollo/client";
+import { Field, FieldContainer, FinalFormSwitch } from "@comet/admin";
+import { Delete, MoreVertical, OpenNewTab, Video } from "@comet/admin-icons";
 import {
     AdminComponentButton,
     AdminComponentPaper,
@@ -11,12 +11,14 @@ import {
     createBlockSkeleton,
     useAdminComponentPaper,
 } from "@comet/blocks-admin";
-import { Box, Divider, Grid, Typography } from "@mui/material";
+import { Box, Divider, Grid, IconButton, ListItemIcon, Menu, MenuItem, Typography } from "@mui/material";
 import { deepClone } from "@mui/x-data-grid/utils/utils";
 import * as React from "react";
 import { FormattedMessage } from "react-intl";
 
 import { DamVideoBlockData, DamVideoBlockInput } from "../blocks.generated";
+import { useContentScope } from "../contentScope/Provider";
+import { useDependenciesConfig } from "../dependencies/DependenciesConfig";
 import { DamPathLazy } from "../form/file/DamPathLazy";
 import { FileField } from "../form/file/FileField";
 import { CmsBlockContext } from "./CmsBlockContextProvider";
@@ -112,7 +114,17 @@ export const DamVideoBlock: BlockInterface<DamVideoBlockData, State, DamVideoBlo
     definesOwnPadding: true,
 
     AdminComponent: ({ state, updateState }) => {
+        const [anchorEl, setAnchorEl] = React.useState<HTMLButtonElement | null>(null);
         const isInPaper = useAdminComponentPaper();
+        const contentScope = useContentScope();
+        const apolloClient = useApolloClient();
+        const dependencyMap = useDependenciesConfig();
+
+        const showMenu = Boolean(dependencyMap["DamFile"]);
+
+        const handleMenuClose = () => {
+            setAnchorEl(null);
+        };
 
         return (
             <Box padding={isInPaper ? 3 : 0} pb={0}>
@@ -132,7 +144,7 @@ export const DamVideoBlock: BlockInterface<DamVideoBlockData, State, DamVideoBlo
                     initialValues={state}
                 >
                     {state.damFile ? (
-                        <FieldContainer label={<FormattedMessage {...messages.file} />} fullWidth>
+                        <FieldContainer fullWidth>
                             <AdminComponentPaper disablePadding>
                                 <Box padding={3}>
                                     <Grid container alignItems="center" spacing={3}>
@@ -146,22 +158,50 @@ export const DamVideoBlock: BlockInterface<DamVideoBlockData, State, DamVideoBlo
                                                 <DamPathLazy fileId={state.damFile.id} />
                                             </Typography>
                                         </Grid>
+                                        {showMenu && (
+                                            <Grid item>
+                                                <IconButton
+                                                    onMouseDown={(event) => event.stopPropagation()}
+                                                    onClick={(event) => {
+                                                        event.stopPropagation();
+                                                        setAnchorEl(event.currentTarget);
+                                                    }}
+                                                    size="large"
+                                                >
+                                                    <MoreVertical />
+                                                </IconButton>
+                                            </Grid>
+                                        )}
                                     </Grid>
                                 </Box>
                                 <Divider />
                                 <AdminComponentButton startIcon={<Delete />} onClick={() => updateState({ damFile: undefined })}>
                                     <FormattedMessage id="comet.blocks.image.empty" defaultMessage="Empty" />
                                 </AdminComponentButton>
+                                {showMenu && (
+                                    <Menu anchorEl={anchorEl} keepMounted open={Boolean(anchorEl)} onClose={handleMenuClose}>
+                                        {dependencyMap["DamFile"] && state.damFile?.id && (
+                                            <MenuItem
+                                                onClick={async () => {
+                                                    // id is checked three lines above
+                                                    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                                                    const path = await dependencyMap["DamFile"].resolvePath({ apolloClient, id: state.damFile!.id });
+                                                    const url = contentScope.match.url + path;
+                                                    window.open(url, "_blank");
+                                                }}
+                                            >
+                                                <ListItemIcon>
+                                                    <OpenNewTab />
+                                                </ListItemIcon>
+                                                <FormattedMessage id="comet.blocks.image.openInDam" defaultMessage="Open in DAM" />
+                                            </MenuItem>
+                                        )}
+                                    </Menu>
+                                )}
                             </AdminComponentPaper>
                         </FieldContainer>
                     ) : (
-                        <Field
-                            name="damFile"
-                            label={<FormattedMessage {...messages.file} />}
-                            component={FileField}
-                            fullWidth
-                            allowedMimetypes={["video/mp4"]}
-                        />
+                        <Field name="damFile" component={FileField} fullWidth allowedMimetypes={["video/mp4"]} />
                     )}
                     <Field
                         type="checkbox"