+++ packages/api/blocks-api/src/blocks/youtube-video.block.ts
@@ -1,4 +1,4 @@
-import { IsBoolean, IsEnum, IsOptional, IsString } from "class-validator";
+import { IsBoolean, IsEnum, IsOptional, IsString, Matches } from "class-validator";
 
 import { BlockData, BlockDataInterface, BlockInput, createBlock, inputToData } from "./block";
 import { BlockField } from "./decorators/field";
@@ -29,6 +29,10 @@ class YouTubeVideoBlockInput extends BlockInput {
     @IsOptional()
     @IsString()
     @BlockField({ nullable: true })
+    // regex from https://stackoverflow.com/a/51870158
+    @Matches(
+        /(https?:\/\/)?(((m|www)\.)?(youtube(-nocookie)?|youtube.googleapis)\.com.*(v\/|v=|vi=|vi\/|e\/|embed\/|user\/.*\/u\/\d+\/)|youtu\.be\/)([_0-9a-zA-Z-]+)/,
+    )
     youtubeIdentifier?: string;
 
     @IsEnum(AspectRatio)