+++ packages/site/cms-site/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @comet/cms-site
 
+## 6.14.1
+
+### Patch Changes
+
+-   84a25adeb: Remove `<a>` tag from `DamFileDownloadLinkBlock`
+
+    The block incorrectly added the tag which prevents styling it in the application. The tag has been removed to achieve the same behavior as in the other link blocks, e.g. `ExternalLinkBlock`.
+
+## 6.14.0
+
+### Minor Changes
+
+-   73dfb61c9: Add `PhoneLinkBlock` and `EmailLinkBlock`
+
+## 6.13.0
+
+### Minor Changes
+
+-   493cad7e1: Add `DamVideoBlock`
+
 ## 6.12.0
 
 ### Minor Changes